requirements.txt
biopython==1.81
numpy==1.24.2
pysam==0.21.0